added onclick event for deletebuttons + extras

+++ individual_list.php
@@ -12,33 +12,34 @@ $tasks = fetch_tasks($database);
 foreach ($lists as $list) :
     if ($list['id'] === $id) : ?>
         <h2> <?= $list['title']; ?> </h2>
-        <button class="delete">
-            <a href="/app/lists/delete.php?id=<?= $list['id']; ?>">Delete list</a>
-        </button>
     <?php endif ?>
 <?php endforeach ?>
 
 <?php foreach ($tasks as $task) :
-    if ($user_id && $list_id = $id) : ?>
-        <h5> <?= $task['title']; ?> </h5>
+    if ($user_id && $list_id = $id) : ?> <br>
+        <h5>Task: <?= $task['title']; ?> </h5>
         <h6>Deadline: <?= $task['deadline_at']; ?> </h6>
-        <button class="delete">
-            <a href="/app/tasks/delete.php?id=<?= $task['id']; ?>">Delete </a>
-        </button>
         <button class="update">
-            <a href="/individual_task.php?id=<?= $task['id']; ?>">Update </a>
+            <a href="/individual_task.php?id=<?= $task['id']; ?>">Update task</a>
         </button>
+        <button class="delete">
+            <a href="/app/tasks/delete.php?id=<?= $task['id']; ?>" onclick="return confirm('Are you sure you want to remove the task?');">Delete task</a>
+        </button>
+        <br>
     <?php endif ?>
 <?php endforeach ?>
-
+<br>
+<h3>Update list</h3>
 <form action="/app/lists/update.php?id=<?= $id ?>" method="post">
     <div class="mb-3">
         <label for="title"></label>
         <input class="form-control" type="text" name="title" id="title" placeholder="New title" required>
         <small class="form-text">Update the name for your list.</small>
     </div>
-    <button type="submit">Change title</button>
-
+    <button class="update" type="submit">Change title</button>
+    <button class="delete">
+        <a href="/app/lists/delete.php?id=<?= $list['id']; ?>" onclick="return confirm('Are you sure you want to delete the list? It will also remove all attached tasks.');">Delete list</a>
+    </button>
 </form>
 <br> <br>
 <h3>Add task</h3>
@@ -59,7 +60,7 @@ foreach ($lists as $list) :
         <small class="form-text">Add a deadline for your task.</small>
     </div>
 
-    <button type="submit">Add task</button>
+    <button class="submit" type="submit">Add task</button>
 
 </form>
